cleanup: use the correct tize_type of the container in CanopyClustering

+++ NaoTHSoccer/Source/Core/Cognition/Modules/Modeling/SelfLocator/MonteCarloSelfLocator/MonteCarloSelfLocatorSimple.cpp
@@ -180,7 +180,7 @@ void MonteCarloSelfLocatorSimple::execute()
       resampleMH(theSampleSet);
       //resampleMHOld(theSampleSet);
 
-      sensorResetBySensingGoalModel(theSampleSet, theSampleSet.size() - 1);
+      sensorResetBySensingGoalModel(theSampleSet, (int)theSampleSet.size() - 1);
 
       // estimate the state
       canopyClustering.cluster(mhBackendSet);
